Boltzmann code reworked [see details below]. Some naming/include conventions changed: python_options.h is now just options.h. [old options.h has been banished to old/src/].  interface/_objects/objects.py got split up as the new complex code was fairly large-ish, it should still appear exactly the same in the package though.
Details on Boltzmann code:

Reworked the generate function to use tempfile to generate temporary files that we don't have to worry about.

While doing so also reworked the subprocess call to avoid one extra file handle being generated/used: it now sends the input straight to sample rather than indirectly via an input file [it still gets to tell sample what file to put the result in, using the tempfile name!].

While doing that I realized it would be trivial to add support for grabbing multiple boltzmann structures. The user can activate this by setting a complex object's .boltzmann_count property to something >= 1. If given, it tries to get that many structures per call to sample, maxing out at 100 per call. Each time a structure is used it's decremented - so once you use up that count it always does 1 per call after that: the onus is on the user to give it a good estimate of the total count.

TODO: add some code inside interface/_options/options.py : MultistrandOptions.start_state to actually give good guesses to the complex objects as to how many times it'll need to sample them. The Options object has a good guess (# of trajectories) for what this is!