Made changes so that boltzmann structures can have parameters coming from the options object.
Now any Complex that is in a start_state gets some private parameters set when it's put into the start state (or when the Options object parameters change). Currently tracked are: dangles, substrate_type and temperature.

Dangles and Substrate type are now full properties with the decorator, so they will be docstringed better, plus that allows us to correctly update the start_state complexes when they change value. Note that it's not entirely clear what to do with RestingState start states; they don't make sense in start states anyways, so the default behavior may be ok.