builder.py
I think when checking convergence with statespace size,  currTime = builderRate.averageTimeFromInitial()  doesn't have to be executed.  It only has to be executed when checking convergence with MFPT ( This suggestion could help because if the statespace is not yet connected from initial to final state, calling  currTime = builderRate.averageTimeFromInitial()  will lead to error. )