update builder.py
I think when checking convergence with statespace size,  builderRate = BuilderRate(self) and  currTime = builderRate.averageTimeFromInitial() shouldn't be executed. It only has to be executed when checking convergence with MFPT.  
Because the states from initial to final state might not be connected and fattenStateSpace is only called after convergence. So "both"  lines could lead to errors when the state space is not connected.