update builder.py
Files which are opened should be closed. For example, myFile = open( ......     , does not close the file, which some times leads to errors such as "  OSError: [Errno 39] Directory not empty: 'p_statespace/-8721681708268404547'  "  .   I've replaced the " myFile = open(...) " to statesments  like "with open( ... ) as myFile: "

I've also added three lines to inspectionSim which sets the buffer conditions. This is important if fatten state space is used to add new states